Adjust current control logic and refine thresholds.
Updated the PWM trigger logic and increased the proportional gain for the current controller. Changed `g.current.ref` to `int32_t` for extended range and adjusted speed thresholds to improve motor state switching behavior.

+++ Drive/mcc/global.h
@@ -56,7 +56,7 @@ typedef struct _speed{
 }TSpeed;
 
 typedef struct _current{
-    volatile int16_t ref;
+    volatile int32_t ref;
     volatile int32_t value;
     volatile int32_t value_peak;
     volatile int32_t limit;
